% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralmetabolic-package.R
\docType{package}
\name{admiralmetabolic-package}
\alias{admiralmetabolic}
\alias{admiralmetabolic-package}
\title{admiralmetabolic: Metabolism Extension Package for ADaM in 'R' Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam}). The package is an extension package of the 'admiral' package focusing on the metabolism therapeutic area.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralmetabolic/}
  \item \url{https://github.com/pharmaverse/admiralmetabolic}
  \item Report bugs at \url{https://github.com/pharmaverse/admiralmetabolic/issues}
}


Other internal: 
\code{\link{derive_param_ratio}()}
}
\author{
\strong{Maintainer}: Anders Askeland \email{iakd@novonordisk.com} (\href{https://orcid.org/0000-0002-2996-129X}{ORCID})

Authors:
\itemize{
  \item Andrii Yurovskyi \email{andrii.yurovskyi@intego-group.com} (\href{https://orcid.org/0009-0006-8606-0929}{ORCID})
  \item Kathrin Flunkert \email{kathrin.flunkert@roche.com}
  \item Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Shunsuke Goto \email{shunsuke.goto@novartis.com}
  \item Siddhesh Pujari \email{spuj@novonordisk.com} (\href{https://orcid.org/0009-0003-4971-1217}{ORCID})
  \item Sonali Das \email{sxd@novonordisk.com}
  \item Olga Starostecka \email{olga.starostecka@boehringer-ingelheim.com}
  \item Vang Le-Quy \email{vqul@novonordisk.com} (\href{https://orcid.org/0000-0003-2035-8619}{ORCID})
  \item Keita Takahashi \email{keita.takahashi.4.biz@gmail.com} (\href{https://orcid.org/0009-0004-9847-5115}{ORCID})
}

}
\concept{internal}
\keyword{internal}
