% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralmetabolic_adsl}
\alias{admiralmetabolic_adsl}
\title{Subject Level Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 54 columns.
}
\source{
Derived from the \code{ad_adsl_metabolic} template
}
\usage{
admiralmetabolic_adsl
}
\description{
An example subject level analysis dataset
}
\seealso{
Other datasets: 
\code{\link{admiralmetabolic_advs}}
}
\concept{datasets}
\keyword{datasets}
