% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralmetabolic_advs}
\alias{admiralmetabolic_advs}
\title{Vital sings Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 760 rows and 101 columns.
}
\source{
Derived from the \code{advs} template
}
\usage{
admiralmetabolic_advs
}
\description{
An example of vital sings analysis dataset
}
\concept{datasets}
\keyword{datasets}
