% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_bds_params.R
\name{compute_ratio}
\alias{compute_ratio}
\title{Compute ratio}
\usage{
compute_ratio(x, y)
}
\arguments{
\item{x}{Dividend

\emph{Permitted Values:} numeric vector}

\item{y}{Divisor

\emph{Permitted Values:} numeric vector}
}
\value{
The ratio computed as x / y.
}
\description{
Computes ratio as dividend (x) divided by divisor (y)
}
\details{
\code{x} and \code{y} must be of the same length.

If the divisor equals 0, the result will be NA.

Usually this computation function can not be used with \code{\%>\%}.
}
\examples{

# Example of Waist-to-Height Ratio where
# Waist Circumference is 120 cm
# and Height is 163 cm

compute_ratio(x = 120, y = 163)
}
\seealso{
\code{\link[=derive_param_ratio]{derive_param_ratio()}}, \code{\link[=derive_param_waisthip]{derive_param_waisthip()}}, \code{\link[=derive_param_waisthgt]{derive_param_waisthgt()}}
}
