% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adlb_params.R
\name{derive_param_glycstt}
\alias{derive_param_glycstt}
\title{Adds a parameter for Glycemic Status}
\usage{
derive_param_glycstt(
  dataset,
  filter = NULL,
  by_vars,
  order,
  set_values_to = exprs(PARAMCD = "GLYCSTT"),
  hba1c_code = "HBA1CHGB",
  fpg_code = "GLUC",
  get_unit_expr,
  prediabetic_thresholds = rlang::list2(`:=`(!!hba1c_code, c(`mmol/mol` = 39, `\%` =
    5.7)), `:=`(!!fpg_code, c(`mmol/L` = 5.6, `mg/dL` = 100))),
  diabetic_thresholds = rlang::list2(`:=`(!!hba1c_code, c(`mmol/mol` = 48, `\%` = 6.5)),
    `:=`(!!fpg_code, c(`mmol/L` = 7, `mg/dL` = 126)))
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.
\code{PARAMCD}, and \code{AVAL} are expected as well.

The variable specified by \code{by_vars} and \code{order} as well as \code{PARAMCD} must be
a unique key of the input dataset after restricting it by the filter condition
(\code{filter} argument) and to the parameters specified by \code{hba1c_code} and \code{fpg_code}.

\describe{
\item{Default value}{none}
}}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\describe{
\item{Permitted values}{an unquoted condition, e.g., \code{AVISIT == "BASELINE"}}
\item{Default value}{\code{NULL}}
}}

\item{by_vars}{Grouping variables

Only variables specified in \code{by_vars} will be populated
in the newly created records.

\describe{
\item{Permitted values}{list of variables created by \code{exprs()}, e.g., \code{exprs(USUBJID, VISIT)}}
\item{Default value}{none}
}}

\item{order}{Sort order

The observations are ordered by the specified order while looking for
consecutive results fulfilling the criteria to confirm a DIABETIC status.

If an expression is named, e.g., \code{exprs(LBDT = convert_dtc_to_dt(LBDTC), LBSEQ)},
a corresponding variable (\code{LBDT}) is added to the additional dataset is used
for sorting. The variable is not included in the output dataset.

For handling of \code{NA}s in sorting variables see the "Sort Order" section in \code{vignette("generic")}.

\describe{
\item{Permitted values}{list of variables created by \code{exprs()}, e.g., \code{exprs(USUBJID, VISIT)}}
\item{Default value}{none}
}}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{exprs(PARAMCD = "GLYCSTT")} defines the parameter code
for the new parameter.

\describe{
\item{Permitted values}{List of variable-value pairs}
\item{Default value}{\code{exprs(PARAMCD = "GLYCSTT")}}
}}

\item{hba1c_code}{HbA1c parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the HbA1c. It is expected that HbA1c is measured in mmol/mol or \%.

\describe{
\item{Permitted values}{character value}
\item{Default value}{\code{"HBA1CHGB"}}
}}

\item{fpg_code}{Fasting Plasma Glucose parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the Fasting Plasma Glucose. It is expected that Fasting Plasma Glucose
is measured in mmol/L or mg/dL.

\describe{
\item{Permitted values}{character value}
\item{Default value}{\code{"GLUC"}}
}}

\item{get_unit_expr}{An expression providing the unit of the parameter

The result is used to check the units of the input parameters.

\describe{
\item{Permitted values}{An expression which is evaluable in the input dataset
and results in a character value}
\item{Default value}{none}
}}

\item{prediabetic_thresholds, diabetic_thresholds}{Criteria for prediabetic and diabetic statuses

The specified thresholds are used to identify HbA1c and Fasting Plasma Glucose
results that corresponds to a certain glycemic status.

\emph{Diabetic:}
\itemize{
\item HbA1c: >=6.5\% (48 mmol/mol)
\item Fasting Plasma Glucose: >=126 mg/dL (7.0 mmol/L)
}

\emph{Prediabetic:}
\itemize{
\item HbA1c: >=5.7\% (39 mmol/mol) but less than the threshold for \emph{Diabetic}
\item Fasting Plasma Glucose: >=100 mg/dL (5.6 mmol/L)
}

\emph{Normoglycemic:}
\itemize{
\item Any values not fulfilling the criteria above
}

Normally the criteria don’t need to be changed, but in case they do,
it can be done as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{prediabetic_thresholds = list(
  HBA1CHGB = xx.x,
  GLUC = xx.x
),
diabetic_thresholds = list(
  HBA1CHGB = xx.x,
  GLUC = xx.x
)
}\if{html}{\out{</div>}}

where \code{xx.x} is a numeric threshold and names of the elements must match
\code{hba1c_code} and \code{fpg_code}.

\describe{
\item{Permitted values}{List of parameter-value pairs}
\item{Default value}{\code{NULL}}
}}
}
\value{
The input dataset with the new parameter added. Note, a variable will only
be populated in the new parameter rows if it is specified in \code{by_vars}.

Possible resulting values (AVALC) are:
\itemize{
\item NORMOGLYCEMIC
\item PREDIABETIC
\item DIABETIC
}
}
\description{
Derives a new parameter using HbA1c and Fasting Plasma Glucose that represents a glycemic status
each by group (e.g., subject and visit) where the source parameters are available.
}
\details{
Implementation is based on ADA 2024.
}
\examples{
library(tibble)
library(rlang)

adlb <- tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD, ~AVISIT, ~AVISITN, ~AVAL, ~AVALU,

  # Subject 1 - From Normoglycemic at BL to Prediabetic at Week 4 onwards

  "STUDY01", "SUBJ001", "HBA1C", "BASELINE", 0, 5.4, "\%",
  "STUDY01", "SUBJ001", "HBA1C", "WEEK 4", 4, 5.8, "\%",
  "STUDY01", "SUBJ001", "HBA1C", "WEEK 8", 8, 6.1, "\%",
  "STUDY01", "SUBJ001", "HBA1C", "WEEK 12", 12, 6.4, "\%",
  "STUDY01", "SUBJ001", "FPG", "BASELINE", 0, 90, "mg/dL",
  "STUDY01", "SUBJ001", "FPG", "WEEK 4", 4, 110, "mg/dL",
  "STUDY01", "SUBJ001", "FPG", "WEEK 8", 8, 115, "mg/dL",
  "STUDY01", "SUBJ001", "FPG", "WEEK 12", 12, 120, "mg/dL",

  # Subject 2 - From Normoglycemic at BL to Diabetic at Week 12

  "STUDY01", "SUBJ002", "HBA1C", "BASELINE", 0, 5.2, "\%",
  "STUDY01", "SUBJ002", "HBA1C", "WEEK 4", 4, 5.6, "\%",
  "STUDY01", "SUBJ002", "HBA1C", "WEEK 8", 8, 6.4, "\%",
  "STUDY01", "SUBJ002", "HBA1C", "WEEK 12", 12, 7.0, "\%",
  "STUDY01", "SUBJ002", "FPG", "BASELINE", 0, 88, "mg/dL",
  "STUDY01", "SUBJ002", "FPG", "WEEK 4", 4, 100, "mg/dL",
  "STUDY01", "SUBJ002", "FPG", "WEEK 8", 8, 120, "mg/dL",
  "STUDY01", "SUBJ002", "FPG", "WEEK 12", 12, 130, "mg/dL",

  # Subject 3 - From Prediabetic at BL to Diabetic at Week 8 onwards

  "STUDY01", "SUBJ003", "HBA1C", "BASELINE", 0, 6.0, "\%",
  "STUDY01", "SUBJ003", "HBA1C", "WEEK 4", 4, 6.2, "\%",
  "STUDY01", "SUBJ003", "HBA1C", "WEEK 8", 8, 6.7, "\%",
  "STUDY01", "SUBJ003", "HBA1C", "WEEK 12", 12, 7.2, "\%",
  "STUDY01", "SUBJ003", "FPG", "BASELINE", 0, 105, "mg/dL",
  "STUDY01", "SUBJ003", "FPG", "WEEK 4", 4, 112, "mg/dL",
  "STUDY01", "SUBJ003", "FPG", "WEEK 8", 8, 130, "mg/dL",
  "STUDY01", "SUBJ003", "FPG", "WEEK 12", 12, 145, "mg/dL"
)

derive_param_glycstt(
  dataset = adlb,
  by_vars = exprs(STUDYID, USUBJID, AVISITN, AVISIT),
  order = exprs(AVISITN),
  set_values_to = exprs(
    PARAMCD = "GLYCSTT"
  ),
  hba1c_code = "HBA1C",
  fpg_code = "FPG",
  get_unit_expr = AVALU
)
}
\concept{der_prm_adlb}
\keyword{der_prm_adlb}
