% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_metabolic}
\alias{dm_metabolic}
\title{Demographic dataset for \code{{admiralmetabolic}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 25 columns.
}
\source{
Constructed using \code{dm} from the \code{{pharmaversesdtm}} package
}
\usage{
dm_metabolic
}
\description{
An example Demographic SDTM dataset for metabolic studies.
}
\keyword{datasets}
