% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qs_metabolic}
\alias{qs_metabolic}
\title{Example questionnaire dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 966 rows and 19 columns.
}
\source{
Constructed using \code{dm_metabolic} from \code{{admiralmetabolic}} package
}
\usage{
qs_metabolic
}
\description{
An example questionnaire SDTM dataset (containing Control of Eating Questionnaire, COEQ, test data) for metabolic studies. Note that University
of Leeds are the copyright holders of the CoEQ and the test data included within \code{{admiralmetabolic}}is for not-for-profit use only. Any persons
or companies wanting to use the CoEQ should request a license to do so from the following
\href{https://licensing.leeds.ac.uk/product/control-of-eating-questionnaire-coeq}{link}.
}
\keyword{datasets}
