% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_advs_params.R
\name{unit-conversion}
\alias{unit-conversion}
\alias{get_conv_factor}
\alias{get_conv_factors_all}
\title{Unit conversion}
\usage{
get_conv_factor(from_unit, to_unit)

get_conv_factors_all()
}
\description{
\code{get_conv_factor()} extracts a conversion factor for a pair of units.
Fails with error if units are not supported/convertible.

\code{get_conv_factors_all()} returns all conversion factors supported.

\strong{Note:} Conversion factors come from unit definitions as per CDISC standards.
}
\keyword{internal}
