% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_metabolic}
\alias{vs_metabolic}
\title{Example Vital Signs dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 719 rows and 24 columns.
}
\usage{
vs_metabolic
}
\description{
An example Vital Signs SDTM dataset for metabolic studies
}
\concept{datasets}
\keyword{datasets}
