% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_metabolic}
\alias{vs_metabolic}
\title{Example vital signs dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 719 rows and 24 columns.
}
\source{
Constructed using \code{vs} from the \code{{pharmaversesdtm}} package and \code{dm_metabolic} from \code{{admiralmetabolic}} package
}
\usage{
vs_metabolic
}
\description{
An example Vital Signs SDTM dataset for metabolic studies.
}
\keyword{datasets}
