% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_metabolic}
\alias{vs_metabolic}
\title{Vital signs Dataset-metabolic}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 678 rows and 24 columns.
}
\usage{
vs_metabolic
}
\description{
An updated SDTM VS dataset for obese patients
}
\concept{datasets}
\keyword{datasets}
